refactor : "Chat 관련 파일 및 클래스명 변경"
ChatRoomController → ChatController
ChatRoomScreen → ChatScreen
date_utils → app_date_utils

Resolve: #109
