+++ lib/presentation/post/widgets/post_item.dart
@@ -6,7 +6,7 @@ import 'package:ondo/core/design_system/app_icon.dart';
 import 'package:ondo/core/design_system/app_layout.dart';
 import 'package:ondo/core/design_system/app_text_styles.dart';
 import 'package:ondo/core/design_system/components/custom_icon_button.dart';
-import 'package:ondo/core/utils/date_utils.dart';
+import 'package:ondo/core/utils/app_date_utils.dart';
 import 'package:ondo/presentation/post/controllers/post_controller.dart';
 
 class PostItem extends GetView<PostController> {
